Fix Vulkan yes OpenGL no builds
The format conversion relies on the QOpenGLTexture header that has its
body ifdefed out with -no-opengl. However, Qt, in a very forward looking
manner, allows having Vulkan support even when OpenGL is disabled.

Assuming that the format conversion will not be used in -no-opengl
builds, put the entire function inside a #if QT_CONFIG(opengl).

Change-Id: I772e12129729d69b81159669d239ea3945a42e5a
Reviewed-by: Johan Helsing <johan.helsing@qt.io>