Improve performance when loading application fonts
Since it was added in Qt 4.2, addAppFont() has been written
to first register the font, then immediately invalidate the
font database and later reload the font again the next time
the db is used. This caused all application fonts to be
reloaded *at least* once, an operation which can be quite heavy
for large fonts, such as CJK fonts. If an application loaded
multiple fonts at different stages of execution, you could
end up loading the same fonts a large number of times, since
all application fonts would be reregistered every time a new
one was added.

When calling removeApplicationFont(), this is okay-ish, since
we need to remove all traces of the font from the platform
database and clearing the whole thing is a convenient way
of making sure there is nothing left. There might be more
efficient ways of doing this, but unloading fonts is not
a common operation, so it is fine to keep this behavior there.

This change removes the invalidation of the font database from
addAppFont(), since this should not be necessary as long as we
are adding fonts. It also removes the reregisterAppFonts flag,
which was a bit of a convoluted way of saying that the database
had been invalidated and needed repopulating. Instead, we use the
same mechanism as for repopulating the system database: We just
check if it is currently empty, which means it has been
invalidated and the application fonts have to be reregistered.

It does not touch the logic in qt_cleanupFontDatabase(), which
is kind of broken (it will leave application fonts in memory
and reregister them if the application continues to run). But
this is only actually called during shutdown (from application
destructor).

[ChangeLog][QtGui][Text] Fixed an issue where application fonts
would be parsed multiple times, causing some unnecessary overhead
when during application startup.

Task-number: QTBUG-76239
Change-Id: Idfb62f73133b55e0909bb398631c8e762442e95b
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>