[androidcontentfileengine] Catch SecurityException during file opening
resolver.openFileDescriptor throws a SecurityException when the caller doesn't have sufficient permissions to access the URL. This will make an application crash.

We should handle this case gracefully and report it to the user. I'm unsure if there is a better way to do this than simply logging the exception.

Change-Id: I36bfbd3e398c8176e57042a27740aa15bce63a7c
Reviewed-by: BogDan Vatra <bogdan@kdab.com>