QCoreWlanEngine: port away from Java-style iterators (and other fixes)
The main goal of this patch was to port away from Java-style
iterators, to make QtBase QT_NO_JAVA_STYLE_ITERATORS-clean. And this
the patch achieves.

But I couldn't resist a few drive-by fixes, too, to wit:

- Use qDeleteAll() instead of while(!isEmpty()) delete takeFirst()

- Use QMap::last() instead of iterating to the end, remembering the
  last-seen value

- Use QMap::contains() instead of QMap::keys().contains()

- Use qExchange() instead of copy+clear

- Make some functions const (requires the mutex member to be marked as
  mutable, which is common for mutex members)

I am almost certain that getSsidFromNetworkName() cannot work correctly.
But this patch does not attempt to change the algorithm.

Change-Id: Ifa04d7837bdc0837036c3a7a73f8c51f4e681f42
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>