Fix 'make check' for android-mingw32
g_options.buildPath did not have native separators. This caused
INSTALL_ROOT to not have native path separators, which again made
'mingw32-make check' fail to deploy the .apk. (The actual failure was that
mkdir failed because it did not accept paths with forward-slashes, so the
install_target rule failed. For the record, the command of that rule had
the following call to mkdir:

mkdir $(INSTALL_ROOT:@msyshack@%=%)\libs\x86

Change-Id: Id792c36986b52a527546d48aa9f7d9587e7a18d9
Reviewed-by: BogDan Vatra <bogdan@kdab.com>
 
 
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>